consolidate home page resolvers
